initial commit!!
Read the code! The first issue I would like to fix is the trails not
showing. It worked but wanted to make a diferent trail for each one of
the cats and it stoped working. Maybe because I added the
xxxxxTrail.draw(mouseX, mouseY); in a plase I should not.